backend/pyproject.toml
[tool.poetry]
name = "social_network"
version = "0.1.0"
description = "Social network"
authors = ["vladyslavyaloveha <vladyslavyaloveha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "django-drf"
django-environ = "^0.10.0"
structlog = "^22.3.0"
django-admin-interface = "^0.24.2"
django-structlog = "^5.0.1"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.1"
psycopg2-binary = "^2.9.5"
django-filter = "^23.1"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular-sidecar = "^2023.4.1"
dj-rest-auth = "^3.0.0"
django-allauth = "^0.54.0"

[tool.poetry.dev-dependencies]
Django = "^4.1.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
